+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/configuration/JsonLdConfiguration.java
@@ -54,14 +54,16 @@ public class JsonLdConfiguration {
     public static final String NAMESPACE_DCT = "http://purl.org/dc/terms/";
     public static final String JSON_LD_OBJECT_MAPPER = "jsonLdObjectMapper";
     public static final String NAMESPACE_CATENAX_POLICY = "https://w3id.org/catenax/policy/";
+    public static final String NAMESPACE_CX_TAXONOMY = "https://w3id.org/catenax/taxonomy#";
+    public static final String NAMESPACE_CX_ONTOLOGY = "https://w3id.org/catenax/ontology/common#";
 
     @Bean /* package */ TitaniumJsonLd titaniumJsonLd(final Monitor monitor) {
         final TitaniumJsonLd titaniumJsonLd = new TitaniumJsonLd(monitor);
-        titaniumJsonLd.registerNamespace("odrl", JsonLdConfiguration.NAMESPACE_ODRL);
+        titaniumJsonLd.registerNamespace("odrl", NAMESPACE_ODRL);
         titaniumJsonLd.registerNamespace("dct", NAMESPACE_DCT);
         titaniumJsonLd.registerNamespace("tx", NAMESPACE_TRACTUSX);
         titaniumJsonLd.registerNamespace("edc", NAMESPACE_EDC);
-        titaniumJsonLd.registerNamespace("dcat", JsonLdConfiguration.NAMESPACE_DCAT);
+        titaniumJsonLd.registerNamespace("dcat", NAMESPACE_DCAT);
         titaniumJsonLd.registerNamespace("dspace", NAMESPACE_DSPACE);
         titaniumJsonLd.registerNamespace("cx-policy", NAMESPACE_CATENAX_POLICY);
         return titaniumJsonLd;